🔧 Refactor: Align route structure to Context OS (blocks, briefs, tasks, creations)
