--- web/app/library/layout.tsx
@@ -1,7 +1,0 @@
-import Shell from "@/components/layouts/Shell";
-
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <Shell>{children}</Shell>;
-}
-
